+++ pannotia/src/tiles/routing_only.rs
@@ -44,7 +44,7 @@ impl FieldPositionCalculator for OMUXRef {
         assert!(self.0 < 16, "output index out of range");
         TileRelativeBitPos {
             x: 15,
-            y: self.0 as u32 * 4 + if self.0 >= 8 { 4 } else { 0 },
+            y: self.0 * 4 + if self.0 >= 8 { 4 } else { 0 },
         }
     }
 }
